+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
+
+version: 0.0.1
+Added very Minor features to this package
+Added Debounce TextField
+Reusable extensions